Move `TestRepository` into `lib/`
This allows it to be configured as a repository for the app in the test
environment going forwards, and lets it be more easily shared between
the app and the sync worker.
